Use just the hostname as MACHINE value in .authinfo.gpg
That way the same key is used for the username as well as the token,
which should help avoid confusion.  Being more specific when looking
up the token doesn't make much sense.  If multiple Github Enterprise
instances exist, then we would expect them to run on different hosts.
And if there are multiple instances, then the same person might have
different names on those instances, which we never supported anyway.
It is also highly unlikely that an instance runs on the same host as
another service that the user might want to authenticate with.