ghub-request: encode METHOD
Caller might represent the method as a symbol and use `symbol-name'
before calling this function, which returns a multibyte string.

That is later concatenated with other parts of the request and if
parts of that contain characters that have to be represented with
bytes in multibyte strings, then we end up with a multibyte string
that actually contains characters that are represented using multiple
bytes.

Because HTTP expects unibyte characters `url-http-create-request'
checks whether `string-bytes' and `length' are equal and in this
scenario they would not be, resulting in a "Multibyte text in HTTP
request" error.

Fixes #35.