ghub--package-scopes: check for missing token later
Previously the first prompt was asking the for a username.  That
prompt is under the control of `url-basic-auth' and provides no
context.

Now the long prompt comes first and if a token by the same name
as the one we are about to create already exists on Github, then
the user is asked about that using an additional prompt.