ghub-request: Allow limiting how many pages are fetched
This is a breaking change - the behavior changes for existing callers
that provided a natural number as a "non-nil value".  To avoid similar
changes in behavior, start raising an error for any value that is not
a boolean or a natural value.  Existing callers might have to be
changed to use (and ... t), which is another breaking change, but
better to get it over with.