ghub--read-json-payload: Signal an error on read error
Do not signal an error if STATUS is 500 because Github always returns
html in that case.

Also remove `ghub--handle-invalid-response', which was previously used
to read such annoying html responses.  We need that anymore because we
now return nil in such cases, so we don't need to construct an error
plist.