ghub--handle-response-error: Use dedicated error symbols and data
Until recently we used dedicated error symbols, but when we started
to build our handling on top of `url's, we dropped that.  Partially
restore that: for HTTP errors use `ghub-http-error' and for other
errors originating from `url' use `ghub-error'.  Unexpected errors
are raised as is.

When signaling an `ghub-http-errors', then append the payload to the
error data, because that contains contains information that might be
useful for debugging.