Ensure url-do-setup has been called before binding variable it sets
When this function has already been called, then it is a
no-op.  Otherwise it sets `url-registered-auth-schemes' among
other things.  If we didn't ensure that it has been run, then
`url-retrieve-synchronously' would do it, which would cause
the value that we let-bind in `ghub--retrieve' to be overwritten,
and the "default" value to be lost outside the let-binding.

Closes #55.