ghub--retrieve: Bind url-show-message to nil
This causes messages such as "Contacting host: api.github.com:443"
not to be shown.  The problem with such messages is that they usually
replace another more informative message, which the author wanted the
user to see but which the user had no chance of seeing because it was
only visible for a millisecond or two.