Don't attempt to retrieve token before we have access
We're facing a bit of a chicken and egg problem here. This function
is called during the flow where we realize that we don't have an API
token.  Because we don't have an API token we want to generate one.
We don't want to generate a new one if we already have one though,
so we try to interact with the API to determine whether we have one,
which we can't do because we don't have a token.

I think it would be better to not trigger the 2FA errorback function
here by augmenting `ghub--get-token-id` to somehow ignore the 2FA
error just for this flow. But I don't have a good instinct as to how
that would fit with the rest of the library.

Fixes #65.