Remove "/v3" from URL when using GraphQL
In the early days I concluded that one base URL was enough for both
the REST and the GraphQL APIs based on the URLs used for github.com:

  REST:    https://api.github.com
  GraphQL: https://api.github.com/graphql

but it turned out not all Github Enterprise instances follow that
format, some instead use:

  REST:    HOST/v3
  GraphQL: HOST/graphql

In such cases "the" base URL (aka the REST URL) is "HOST/v3" and we
have to remove the "/v3" before appending "/graphql".  It's a hack,
but also too late to do it differently.

Closes https://github.com/magit/forge/issues/174.