ghub-fetch-repository: Fetch reviewThreads, baseRefOid/headRefOid
PullRequestReviewThread object is a threaded list of review comments
for a given pull request.

headRefOid and baseRefOid are the respective head and base SHA
reference for a given pull request.

The GitHub GraphQL API has limitations, Individual calls cannot
request more than 500,000 total nodes.  Thus in order to fetch
reviewThreads from a pull-request, we need to fetch these data
in a dedicated function: `ghub-fetch-review-threads'.

Otherwise we will have this kind of error message: "By the time
this query traverses to the comments connection, it is requesting
up to 1,000,000 possible nodes which exceeds the maximum limit of
500,000."

Source:
https://docs.github.com/en/free-pro-team@latest/graphql/overview/resource-limitations
https://docs.github.com/en/free-pro-team@latest/graphql/reference/objects#pullrequestreviewthread

Signed-off-by: Julien Masson <massonju.eseo@gmail.com>