Advice url-http-handle-authentication with a bugfix
Upon receiving a 401 or 409 `url-http-parse-headers' calls
`url-http-handle-authentication', which then makes another
request, which now provides the authentication information.

Unfortunately that function does not react correctly when the previous
request provided authentication information but the server rejected it.
In that case it would be appropriate to signal a `http' error.

On Emacs releases before 27.1 it instead makes another request, which
can result in either an infinite loop if all the information required
for a request using basic authentication is already available, until
the server refuses to comply; or, as in the case of `ghub' it results
in the user being asked for their username and/or password, which is
very confusing.

On Emacs 27.1 it does signal an error but without using the mechanism
intended to do so, resulting in an error that does not contain all the
information needed to properly deal with it.

See https://github.com/magit/forge/issues/293.