ghub-response-link-relations: Fix for GitLab
They changed the name of the header from `Link' to `link', or maybe
they always did the latter (though I expect this issue would have been
reported a long time ago, if that were the case).

To this day, the section [1] where they documents this header, is name
"Pagination `Link' header" (with `Link' using code syntax), which is
misleading, especially consider that HTTP headers traditionally are
capitalized [2], but in their API it is not (while some others are).

[1]: https://docs.gitlab.com/ee/api/rest/index.html#pagination-link-header
[2]: https://www.iana.org/assignments/http-fields/http-fields.xhtml#field-names