No longer prepare GraphQL queries prematurely
`ghub--graphql-vacuum' calls
  `ghub--graphql-retrieve' calls
    `ghub--graphql-prepare-query'
      - calls `ghub--graphql-narrow-query'
        Which doesn't do anything for this first request because
        LINEAGE is nil.
      - and prepares the QUERY for `gsexp-encode'

All `ghub-fetch-*' functions (except `ghub-fetch-repository') first
need to narrow the value of the `ghub-fetch-repository' variable to the
part they are interested in.  They used `ghub--graphql-narrow-query',
which means that that function was run twice before even making the
first request.  Until now that wasn't a big issue, the query was not
modified the second time around, but it was a bit wasteful.

The next commit adds a PAGINATE argument to `ghub--graphql-vacuum',
which stores that in a new PAGINATE field slot of the request object.
`ghub--graphql-retrieve' passes the slots value to
`ghub--graphql-prepare-query', which takes it into account when
preparing the query.

Part of what `ghub--graphql-prepare-query' does to the query is to
encode pagination.  This is done in a way that loses information.
[(:edges E)] becomes [(first F)))], where F depends on E and the
pagination information from the request object.  Once F is stored
in the query, E is lost, so calling this function again (with new
pagination information from the object), cannot change the outcome
in the query anymore.

That is very easy to avoid: use `ghub--graphql-narrow-query' for
the early call, avoiding the unnecessary, destructive preparation.