ghub--graphql-vacuum: Add PAGINATE argument
Add the same argument to internal caller of that function.  Store
the value in the PAGINATE slot of the request object, and finally
put it in action in `ghub--graphql-prepare-query'.

Previously a caller had to pass this information by let-binding
`ghub-graphql-items-per-request', if it did not want to use the
global default value.

Also see previous commit.