ghub--graphql-prepare-query: Use smallest pagination setting
Previously if pagination was explicitly specified for a certain edge,
then that was used unconditionally, even if it happened to be larger
than the "maximum" specified for "all" edges, using the PAGINATE
argument or `ghub-graphql-items-per-request' variable.

Now, if a caller specifies PAGINATE or the user lowers
`ghub-graphql-items-per-request', they can be sure that that is always
respected.