Handle timeouts properly
No longer use a "BUG" error, instead handle it as other kinds of errors
are, by either passing it to the error-/callback, signaling an error, or
simply returning nil (depending on what the caller asked us to do).

(`ghub-handle-response-headers' actually still has code that would
signal a "BUG" error, but it is expected that this can no longer happen.
We keep it around in case it does still happen, in which case we would
really want to know about it.  Also note that if a caller failed to make
the response buffer current before calling this function, then we would
hit this too (but that would be a new bug, and is not the reason why we
keep this check around).