ghub--read-json-payload: On parse error display response buffer
Github promised to deliver json but sometimes returns html instead.
Display the response buffer when that happens.  Previously we only
displayed part of it in an error message, which risks losing track
of some relevant information.

If this turns out to cause more issues than it helps solving, we
might have to only do this if `ghub-debug' is enabled globally.