Deprecate {gogs,gtea,buck}-VERB functions
These symbols (and the libraries in which they are defined) do not
use the prefix of this package.  We are squatting on prefixes that
could otherwise be used by other packages.

Additionally adding the package prefix to these symbols and
libraries, as in `ghub-gogs-get', would defeat the whole purpose
of these functions, so instead these functions and libraries will
eventually have to be removed.

For the time being do not deprecate `glab-VERB' and
`FORGE-repository-id', but these are also subject to removal.