ghub--graphql-vacuum: Allow using callback with synchronous requests
Add argument SYNCHRONOUS as a the new mechanism to specify that the
request is synchronous.  Passing `synchronous' as CALLBACK no longer
works.  The `ghub--req' struct needs a new slot to pass along whether
the request is synchronous, since we can no longer check the value of
the CALLBACK slot.

Start using `ghub--graphql-handle-success' for synchronous requests
as well.

For REST requests CALLBACK and ERRORBACK are still used to determine
whether to make a synchronous request or not.  Start making that
determination in `ghub-request' instead of `ghub--receive' (which
is necessary because the latter is also used for GraphQL requests).