ghub--graphql-vacuum: Add NOERROR argument
If ERRORBACK is non-nil, this new argument has no effect.  Otherwise
non-nil NOERROR means that CALLBACK is called, if non-nil, or else
that it is silently ignored.

If the special value `pp' is used as CALLBACK, t is used for NOERROR,
causing `ghub--graphql-pp-response' to be used for errors as well
(unless ERRORBACK is non-nil).