ghub--graphql-handle-response: Maybe use success handler despite errors
If the response contains `data' in addition to `errors' and
`ghub--query's NOERROR argument was non-nil, then do not treat
this as an error.  Call `ghub--graphql-handle-success' instead
of `ghub--graphql-handle-failure'.

This is used by the `forge--pull-topic' method, which only knows the
topic number but not its type and thus attempts to fetch three topics,
one for each topic type.  This always results in two error, but we want
to continue fetching data for the topic of the correct type.  If we did
instead let `ghub--graphql-handle-failure' call CALLBACK, then
unresolved edges would remain in the processed data.