ghub--auth-source-get: Do not add :max 1 to search spec
Adding this is neither necessary nor sufficient to guarantee only one
item is returned:

> :max N means to try to return at most N items (defaults to 1).
> More than N items may be returned, depending on the search and
> the backend.

As before, use the first, and hopefully only, item.  What changes is
only that we no longer make it explicit that we would prefer to only
receive one item.

Alternatively we could have started to use the extended spec not only
when searching but also when invalidating the result.  Note that while
`auth-source-search' ignores `:max' when searching, it doesn't do so
when caching, so when we remove from the cache, we have to include
`:max' because it is part of the key used in the cache hashtable.

Closes #180.